% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplitude_altimetrica.R
\name{amplitude_alimetrica}
\alias{amplitude_alimetrica}
\title{Amplitude Altimétrica}
\usage{
amplitude_alimetrica(x)
}
\arguments{
\item{x}{Entrada do arquivo matricial do MDE.}

\item{escala}{Númerico. Efetua a escala espacial.}
}
\description{
\strong Cálcula a Amplitude Altimétrica.
}
