% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_topo_morro.R
\name{app_topo_morro}
\alias{app_topo_morro}
\title{APP Topo de Morros}
\usage{
app_topo_morro(x)
}
\arguments{
\item{x}{Entrada do arquivo matricial MDE.}
}
\description{
\strong{APP Topo de Morros}
}
\examples{
\dontrun{
library(terra)
library(rgeoamb)
dem <- terra::rast(system.file('ex/elev.tif', package = 'terra'))
app_tm <- rgeoamb::app_topo_morro(dem)
plot(app_tm)
}
}
