% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_perimeter.R
\name{area_perimeter}
\alias{area_perimeter}
\title{Area and Perimeter}
\usage{
area_perimeter(aoi)
}
\arguments{
\item{aoi}{Area of interest. Must be an `sf` object.}
}
\description{
Calculates the area in square meters, square kilometers, and hectares, as well as the perimeter.
}
\examples{
aoi <- sf::read_sf(system.file('ex/aoi.shp', package = 'sf'))|>
       subset(ex == 'area of interest')
ap <- rgeoamb::area_perimeter(aoi)
print(ap)
}
