% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_perimetro.R
\name{area_perimetro}
\alias{area_perimetro}
\title{Área e Perímetro}
\usage{
area_perimetro(aoi)
}
\arguments{
\item{aoi}{Área de interesse, deve ser um objeto sf.}
}
\description{
\strong Cálcula a área em metros e quilomêtros quadrados, hectares e perímetro.
}
