% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_uso_restrito.R
\name{area_uso_restrito}
\alias{area_uso_restrito}
\title{Área de Uso Restrito}
\usage{
area_uso_restrito(x)
}
\arguments{
\item{x}{Entrada do arquivo matricial MDE.}
}
\description{
\strong{Área de Uso Restrito}
}
\examples{
\dontrun{
library(terra)
library(rgeoamb)
dem <- terra::rast(system.file('ex/elev.tif', package = 'terra'))
uso_restrito <- rgeoamb::area_uso_restrito(dem)
plot(uso_restrito)
}
}
