% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas_verdes.R
\name{areas_verdes}
\alias{areas_verdes}
\title{Detecção de Áreas Verdes}
\usage{
areas_verdes(nir, red, amm, l1, l2, l3)
}
\arguments{
\item{nir}{Entrada do arquivo matricial da banda do infravermelho (nir).}

\item{red}{Entrada do arquivo matricial da banda do vermelho (nir).}

\item{amm}{Númerico. Depende da função Área Mínima Mapeável.}

\item{l1}{Númerico. Primeiro intervalo para determinar a extração das áreas verdes.}

\item{l2}{Númerico. Segundo Intervalo para determinar a extração das áreas verdes.}

\item{l3}{Númerico. Intervalos entre os limiares para determinar a extração das áreas verdes.}
}
\description{
\strong{Detecção de Áreas Verdes}
}
\examples{
\dontrun{
library(terra)
library(rgeoamb)
nir <- terra::rast(system.file('ex/nir.tif', package = 'terra'))
red <- terra::rast(system.file('ex/red.tif', package = 'terra'))
av <- rgeoamb::mascara_vegetacao(nir, red, 5, 0.33, 0.40, 0.01)
plot(av)
}
}
