% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stac_services.R
\name{download_stac_services}
\alias{download_stac_services}
\title{Download STAC Image Collections from Planetary Computer}
\usage{
download_stac_services(
  collections,
  date_time,
  aoi,
  cloud_cover,
  bands,
  output_dir
)
}
\arguments{
\item{collections}{Character. Name of the STAC collection(s) (e.g., `"sentinel-2-l2a"` or `"landsat-c2-l2"`).}

\item{date_time}{Character. ISO 8601 formatted date or date-time range (e.g., `"2023-01-01/2023-12-31"`).}

\item{aoi}{An `sf` object representing the spatial area of interest.}

\item{cloud_cover}{Numeric. Maximum allowed cloud cover percentage (0 to 100).}

\item{bands}{Character vector. Names of the asset/band names to select (e.g., `c("B04", "B08")`).}

\item{output_dir}{Character. Directory path where raster files will be saved.}
}
\description{
Query, filter, sign, and download satellite assets from Microsoft's
Planetary Computer STAC API using a spatial area of interest.
}
\examples{
\dontrun{
library(pacman)
p_load(sf, rgeoamb)
aoi <- sf::read_sf('ex/aoi.shp')
data <- rgeoamb::download_stac_services('landsat-c2-l2', '2021-01-01/2021-12-31', aoi, 10, c('nir08', 'red'), 'ex/image')
print(data)
}
}
