% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_basins.R
\name{drainage_basins}
\alias{drainage_basins}
\title{Drainage Basins (Bacias Hidrográficas)}
\usage{
drainage_basins(x, proj)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{proj}{Performs geodesic reference frame transformations and projective transformations.}
}
\description{
Identifies drainage basins in DEMs.
}
\examples{
library(terra)
library(rgeoamb)
dem <- terra::rast(system.file('ex/elev.tif', package = 'terra'))
db <- rgeoamb::drainage_basins(dem)
}
