% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_areas.R
\name{green_areas}
\alias{green_areas}
\title{Green Areas}
\usage{
green_areas(nir, red, mma, l1, l2, l3)
}
\arguments{
\item{nir}{Input raster object for the near-infrared (NIR) band input.}

\item{red}{Input raster object for the red band input.}

\item{amm}{Numeric. Minimum Mappable Area constraint.}

\item{t1}{Numeric. First threshold value to extract green areas.}

\item{t2}{Numeric. Second threshold value to extract green areas.}

\item{t3}{Numeric. Step interval between thresholds to extract green areas.}
}
\description{
Detects and extracts green areas from spatial raster data.
}
\examples{
library(terra)
library(rgeoamb)
nir <- terra::rast(system.file('ex/nir.tif', package = 'terra'))
red <- terra::rast(system.file('ex/red.tif', package = 'terra'))
ga <- rgeoamb::green_areas(nir, red, 5, 0.33, 0.40, 0.01)
plot(ga)
}
