% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilltops.R
\name{hilltops}
\alias{hilltops}
\title{Hilltops (APPs Topo de Morros)}
\usage{
hilltops(x, proj)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{proj}{Performs geodesic reference frame transformations and projective transformations.}
}
\description{
Identifies Permanent Preservation Areas (APPs) in hilltops.
}
\examples{
library(terra)
library(rgeoamb)
dem <- terra::rast(system.file('ex/elev.tif', package = 'terra'))
htps <- rgeoamb::hilltops(dem)
plot(htps)
}
