% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse_landcover.R
\name{landuse_landcover}
\alias{landuse_landcover}
\title{Land Use and Land Cover}
\usage{
landuse_landcover(
  x,
  y,
  target_col,
  p,
  drop_cols,
  k_folds,
  method,
  mtry,
  ntree,
  cost,
  preProcess,
  path_var_imp,
  path_prediction
)
}
\arguments{
\item{x}{Input stack raster files.}

\item{y}{Input samples (vector files) in terra or sf package R.}

\item{target_col}{Numeric. Target field of occurrence.}

\item{p}{Numeric. Regarding the number of samples representing the non-occurrence of events, it is suggested, in order to balance the sample set, to use a sample size equal to the number of event occurrences.}

\item{drop_cols}{This allows you to remove information that will not be used in susceptibility modeling.}

\item{k_folds}{Numeric. Number of folds or number of resampling iterations, using cross-validation method.}

\item{method}{Random Forest 'rf' or Support Vector Machine 'svm'}

\item{mtry}{Numeric. Randomly selected predictors}

\item{ntree}{Numeric. Decision tree number}

\item{cost}{Numeric. Constraint violation cost, where 'C' is the regularization parameter.}

\item{preProcess}{Estimates the required parameters for each operation (e.g., 'nzv' for near-zero variance filtering, or c('center', 'scale') for normalization). For details on available pre-processing methods, see the caret package R documentation.}

\item{path_var_imp}{Character. Path to save the machine learning model variable importance in tabular format.}

\item{path_prediction}{Character. Path to save the susceptibility prediction in raster/matrix format.}
}
\description{
Performing land use and land cover mapping through machine learning, based on Filho et al. (2024).
}
\examples{
library(pacman)
p_load(terra, rgeoamb)
rasters <- terra::rast(system.file("ex/stack.tif", package = "terra"))
samples <- terra::vect(system.file("ex/samples.shp", package = "terra"))
lulc <- rgeoamb::landuse_landcover(rasters, samples, "classes", 0.7, -c(1, 2, 10), 4, 'rf', 6, 500, NULL,
                                             "nzv", "ex/var_imp.txt", "ex/predict.tif")
plot(lulc)
}
