% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_reserve.R
\name{legal_reserve}
\alias{legal_reserve}
\title{Legal Reserve (Reserva Legal)}
\usage{
legal_reserve(x, y)
}
\arguments{
\item{x}{Input vector object for the property area.}

\item{y}{Input vector object for the legal reserve.}
}
\description{
Checks if the property area complies with the Brazilian Forest Code.
}
\examples{
library(sf)
library(rgeoamb)
area_imovel <- sf::read_sf(system.file('ex/area_imovel.shp', package = 'sf'))
lr <- sf::read_sf(system.file('ex/lr.tif', package = 'sf'))
plr <- rgeoamb::legal_reserve(property_area, lr)
plot(plr)
}
