% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_relief.R
\name{relative_relief}
\alias{relative_relief}
\title{Relative Relief (Amplitude Altimétrica)}
\usage{
relative_relief(x)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{scale}{Numeric. Number of neighbor cells for multiscalar analysis.}
}
\description{
Calculates Relative Relief in DEMs.
}
\examples{
library(terra)
library(rgeoamb)
dem <- terra::rast(system.file('ex/elev.tif', package = 'terra'))
rr <- rgeoamb::relative_relief(dem)
plot(rr)
}
