% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted_use_area.R
\name{restricted_use_area}
\alias{restricted_use_area}
\title{Restricted Use Area (Área de Uso Restrito)}
\usage{
restricted_use_area(x)
}
\arguments{
\item{x}{Input DEM raster file.}
}
\description{
Detects restricted use area in DEMs.
}
\examples{
library(terra)
library(rgeoamb)
dem <- terra::rast(system.file('ex/elev.tif', package = 'terra'))
restricted_use <- rgeoamb::restricted_use_area(dem)
plot(restricted_use)
}
