% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_bodies.R
\name{water_bodies}
\alias{water_bodies}
\title{Water Bodies (APPs Corpos Hídricos)}
\usage{
water_bodies(green, nir, threshold, buffer)
}
\arguments{
\item{green}{Input raster object for the green band input.}

\item{nir}{Input raster object for the near-infrared (NIR) band input.}

\item{threshold}{Numeric. Threshold used to extract water bodies.}

\item{buffer}{Numeric. Buffer distance (zone of influence).}
}
\description{
Identifies Permanent Preservation Areas (APPs) in water bodies.
}
\examples{
library(terra)
library(rgeoamb)
green <- terra::rast(system.file('ex/green.tif', package = 'terra'))
nir <- terra::rast(system.file('ex/nir.tif', package = 'terra'))
wb <- rgeoamb::water_bodies(green, nir, 0.1)
plot(wb)
}
