% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_mask.R
\name{water_mask}
\alias{water_mask}
\title{Water Mask (Máscara d'água)}
\usage{
water_mask(green, nir, amm, t1, t2, t3)
}
\arguments{
\item{green}{Input raster object for the green band input.}

\item{nir}{Input raster object for the near-infrared (NIR) band input.}

\item{amm}{Numeric. Minimum Mappable Area constraint.}

\item{t1}{Numeric. First threshold value to extract water bodies.}

\item{t2}{Numeric. Second threshold value to extract water bodies.}

\item{t3}{Numeric. Step interval between thresholds to extract water bodies.}
}
\description{
Detects and extracts water bodies (masks) from spatial raster data.
}
\examples{
library(terra)
library(rgeoamb)
green <- terra::rast(system.file('ex/green.tif', package = 'terra'))
nir <- terra::rast(system.file('ex/nir.tif', package = 'terra'))
wm <- rgeoamb::water_mask(green, nir, 5, 0.1, 0.10, 0.1)
plot(wm)
}
