% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_45.R
\name{app_45}
\alias{app_45}
\title{APP > 45°}
\usage{
app_45(x)
}
\arguments{
\item{x}{Input DEM raster file.}
}
\description{
Identifies Permanent Preservation Areas (APPs) in slope angles > 45°.
}
\examples{
library(terra)
library(rgeoamb)
dem <- terra::rast(system.file('ex/elev.tif', package = 'terra'))
app <- rgeoamb::app_45(dem)
plot(app)
}
