% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_corpos_hidricos.R
\name{app_corpos_hidricos}
\alias{app_corpos_hidricos}
\title{APP Corpos Hídricos}
\usage{
app_corpos_hidricos(green, nir, limiar)
}
\arguments{
\item{limiar}{Númerico. Limiar para determinar a extração de corpos hídricos}

\item{x}{Entrada do arquivo matricial da banda verde (green)}

\item{y}{Entrada do arquivo matricial da banda do infravermelho (nir)}

\item{buffer}{Númerico. Área de influência.}
}
\description{
\strong APP Corpos Hídricos
}
\examples{
library(terra)
library(rgeoamb)
green <- terra::rast(system.file('ex/green.tif, package = 'terra'))
nir <- terra::rast(system.file('ex/nir.tif, package = 'terra'))
app_ch <- rgeoamb::app_corpos_hídricos(green, nir, 0.1)
plot(app_ch)
}
