% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_minima_mapeavel.R
\name{area_minima_mapeavel}
\alias{area_minima_mapeavel}
\title{\strong Área Mínima Mapeável}
\usage{
area_minima_mapeavel(escala, res)
}
\arguments{
\item{escala}{Númerico. Escala cartográfica de mapeamento.}

\item{res}{Númerico. Resolução da imagem.}
}
\description{
Calcula a Área Mínima Mapeável
}
\examples{
library(rgeoamb)
amm <- rgeoamb::area_minima_mapeavel(25000, 10)
}
