% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascara_dagua.R
\name{mascara_dagua}
\alias{mascara_dagua}
\title{Máscara d'água}
\usage{
mascara_dagua(green, nir, amm, l1, l2, l3)
}
\arguments{
\item{green}{Entrada do arquivo matricial da banda verde (green).}

\item{nir}{Entrada do arquivo matricial da banda do infravermelho (nir).}

\item{amm}{Númerico. Depende da função Área Mínima Mapeável}

\item{l1}{Númerico. Primeiro intervalo para determinar a extração de corpos hídricos.}

\item{l2}{Númerico. Segundo Intervalo para determinar a extração de corpos hídricos.}

\item{l3}{Númerico. Intervalos entre os limiares para determinar a extração de corpos hídricos.}
}
\description{
\strong{Máscara d'água}
}
\examples{
\dontrun{
library(terra)
library(rgeoamb)
green <- terra::rast(system.file('ex/green.tif', package = 'terra'))
nir <- terra::rast(system.file('ex/nir.tif', package = 'terra'))
wm <- rgeoamb::mascara_dagua(green, nir, 5, 0.1, 0.10, 0.1)
plot(wm)
}
}
