% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascara_vegetacao.R
\name{mascara_vegetacao}
\alias{mascara_vegetacao}
\title{Máscara de Vegetação}
\usage{
mascara_vegetacao(nir, red, amm, l1, l2, l3)
}
\arguments{
\item{nir}{Entrada do arquivo matricial da banda do infravermelho (nir).}

\item{red}{Entrada do arquivo matricial da banda do vermelho (nir).}

\item{amm}{Númerico. Depende da função Área Mínima Mapeável.}

\item{l1}{Númerico. Primeiro intervalo para determinar a extração das áreas vegetadas.}

\item{l2}{Númerico. Segundo Intervalo para determinar a extração das áreas vegetadas.}

\item{l3}{Númerico. Intervalos entre os limiares para determinar a extração das áreas vegetadas.}
}
\description{
\strong{Máscara de Vegetação}
}
\examples{
\dontrun{
library(terra)
library(rgeoamb)
nir <- terra::rast(system.file('ex/nir.tif', package = 'terra'))
red <- terra::rast(system.file('ex/red.tif', package = 'terra'))
vm <- rgeoamb::mascara_vegetacao(nir, red, 5, 0.33, 0.40, 0.01)
plot(vm)
}
}
