% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserva_legal.R
\name{reserva_legal}
\alias{reserva_legal}
\title{Reserva Legal}
\usage{
reserva_legal(x, y)
}
\arguments{
\item{x}{Entrada do arquivo vetorial da área do ímovel.}

\item{y}{Entrada do arquivo vetorial da reserva legal.}
}
\description{
\strong{Cálcula se a área do imóvel condiz com o código florestal.}
}
\examples{
\dontrun{
library(sf)
library(rgeoamb)
area_imovel <- sf::read_sf(system.file('ex/area_imovel.shp', package = 'sf'))
rl <- sf::read_sf(system.file('ex/rl.tif', package = 'sf'))
rlp <- rgeoamb::reserva_legal(area_imovel, rl)
plot(rlp)
}
}
